+++ src/modules/address/repository/address.repository.ts
@@ -1,5 +1,5 @@
 import { EntityRepository } from 'typeorm';
-import { AddressEntity } from '@modules/database/entity/address.entity';
+import { AddressEntity } from '@modules/address/entity/address.entity';
 import { BaseRepository } from 'typeorm-transactional-cls-hooked';
 
 @EntityRepository(AddressEntity)
