+++ src/modules/comment/repository/comment.repository.ts
@@ -1,6 +1,6 @@
 import { EntityRepository } from 'typeorm';
 import { BaseRepository } from 'typeorm-transactional-cls-hooked';
-import { CommentEntity } from '@modules/database/entity/comment.entity';
+import { CommentEntity } from '@modules/comment/entity/comment.entity';
 
 @EntityRepository(CommentEntity)
 export class CommentRepository extends BaseRepository<CommentEntity> {}
