+++ src/modules/menu-item/repository/menu-item.repository.ts
@@ -1,6 +1,6 @@
 import { EntityRepository } from 'typeorm';
 import { BaseRepository } from 'typeorm-transactional-cls-hooked';
-import { MenuItemEntity } from '@modules/database/entity/menu-item.entity';
+import { MenuItemEntity } from '@modules/menu-item/entity/menu-item.entity';
 
 @EntityRepository(MenuItemEntity)
 export class MenuItemRepository extends BaseRepository<MenuItemEntity> {}
