+++ src/modules/order/repository/order.repository.ts
@@ -1,6 +1,6 @@
 import { EntityRepository } from 'typeorm';
 import { BaseRepository } from 'typeorm-transactional-cls-hooked';
-import { OrderEntity } from '@modules/database/entity/order.entity';
+import { OrderEntity } from '@modules/order/entity/order.entity';
 
 @EntityRepository(OrderEntity)
 export class OrderRepository extends BaseRepository<OrderEntity> {}
