+++ src/modules/restaurant/entity/restaurant.entity.ts
@@ -5,11 +5,11 @@ import {
   Entity,
   OneToMany,
 } from 'typeorm';
-import { RowEntity } from '@modules/database/entity/shared/row.entity';
+import { RowEntity } from '@modules/database/entity/row.entity';
 import { validateOrReject } from 'class-validator';
-import { CommentEntity } from '@modules/database/entity/comment.entity';
-import { MenuItemEntity } from '@modules/database/entity/menu-item.entity';
-import { OrderEntity } from '@modules/database/entity/order.entity';
+import { CommentEntity } from '@modules/comment/entity/comment.entity';
+import { MenuItemEntity } from '@modules/menu-item/entity/menu-item.entity';
+import { OrderEntity } from '@modules/order/entity/order.entity';
 
 @Entity('restaurant')
 export class RestaurantEntity extends RowEntity<RestaurantEntity> {
