+++ src/modules/restaurant/repository/restaurant.repository.ts
@@ -1,6 +1,6 @@
 import { EntityRepository } from 'typeorm';
 import { BaseRepository } from 'typeorm-transactional-cls-hooked';
-import { RestaurantEntity } from '@modules/database/entity/restaurant.entity';
+import { RestaurantEntity } from '@modules/restaurant/entity/restaurant.entity';
 
 @EntityRepository(RestaurantEntity)
 export class RestaurantRepository extends BaseRepository<RestaurantEntity> {}
