+++ src/modules/user-role/repository/user-role.repository.ts
@@ -1,6 +1,6 @@
 import { EntityRepository } from 'typeorm';
 import { BaseRepository } from 'typeorm-transactional-cls-hooked';
-import { UserRoleEntity } from '@modules/database/entity/user-role.entity';
+import { UserRoleEntity } from '@modules/user-role/entity/user-role.entity';
 
 @EntityRepository(UserRoleEntity)
 export class UserRoleRepository extends BaseRepository<UserRoleEntity> {}
