+++ src/modules/user/user.module.ts
@@ -1,5 +1,5 @@
-import { UserRepository } from '@modules/database/repository/user.repository';
-import { UserRoleRepository } from '@modules/database/repository/user-role.repository';
+import { UserRepository } from '@modules/user/repository/user.repository';
+import { UserRoleRepository } from '@modules/user-role/repository/user-role.repository';
 import { UserService } from '@modules/user/user.service';
 import { UserController } from '@modules/user/user.controller';
 import { Module } from '@nestjs/common';
